Make application running again
Code got rearranged between lib and main executable in the process. Pomodoro.Application and d-bus service are somewhat related to GTK parts of the code, so more things were moved out from the Application class. But unltimately d-bus Service should be included as part of the lib and have tests.