extension/screenshield: Don't wake up screen on extension start
Fixes #649.

Before this commit, ScreenShieldManager._ontimerStateChanged() was
called on extension startup and before the DBus connection was live,
which could issue an attempt to wake up the screen during extension
start. This is apparently problematic (see #640) particular, because
extension startup can happen during the first screen lock due to recent
changes in GNOME45:
https://discourse.gnome.org/t/extension-disable-called-on-session-mode-change-but-only-for-the-first-time/15832
https://discourse.gnome.org/t/would-a-partial-disable-based-on-extensionstate-disabling-which-only-reverts-patches-be-permissible/17426

Moreover, we should not wake up the screen when we get the initial status
update from DBus.